backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
yfinance==0.2.33
pandas==2.1.4
numpy==1.26.2
requests==2.31.0
google-generativeai==0.3.2
transformers==4.36.2
torch==2.1.2
scikit-learn==1.3.2
python-dotenv==1.0.0
